+++ src/extraction/special/comments.rs
@@ -1,23 +1,43 @@
 use winnow::{
-    ascii::{line_ending, multispace0, multispace1, till_line_ending},
+    ascii::{line_ending, multispace1, till_line_ending},
     combinator::{alt, delimited, repeat, trace},
+    token::take_while,
     BStr, PResult, Parser,
 };
 
-/// A comment extractor. This parser returns the line after the `%` sign.
-fn comment<'de>(input: &mut &'de BStr) -> PResult<&'de [u8]> {
-    trace(
-        "livre-comment",
-        delimited((b'%', multispace0), till_line_ending, line_ending),
-    )
-    .parse_next(input)
+use crate::extraction::Extract;
+
+/// PDF comment, which references into the original data.
+///
+/// The content is borrowed, starts after the first leading `%` and strips leading spaces.
+pub struct Comment<'de>(pub &'de [u8]);
+
+impl<'de> Extract<'de> for Comment<'de> {
+    fn extract(input: &mut &'de BStr) -> PResult<Self> {
+        trace(
+            "livre-comment",
+            delimited(
+                (b'%', take_while(0.., (b' ', b'\t'))),
+                till_line_ending,
+                line_ending,
+            )
+            .map(Comment),
+        )
+        .parse_next(input)
+    }
+
+    fn recognize(input: &mut &'de BStr) -> PResult<&'de [u8]> {
+        (b'%', till_line_ending, line_ending)
+            .take()
+            .parse_next(input)
+    }
 }
 
 /// Parses 0 or more comment/multispace
 pub fn multicomment0<'de>(input: &mut &'de BStr) -> PResult<&'de [u8]> {
     trace(
         "livre-comments-whitespace",
-        repeat::<_, _, (), _, _>(0.., alt((comment, multispace1))).take(),
+        repeat::<_, _, (), _, _>(0.., alt((Comment::recognize, multispace1))).take(),
     )
     .parse_next(input)
 }
@@ -26,7 +46,7 @@ pub fn multicomment0<'de>(input: &mut &'de BStr) -> PResult<&'de [u8]> {
 pub fn multicomment1<'de>(input: &mut &'de BStr) -> PResult<&'de [u8]> {
     trace(
         "livre-comments-whitespace",
-        repeat::<_, _, (), _, _>(1.., alt((comment, multispace1))).take(),
+        repeat::<_, _, (), _, _>(1.., alt((Comment::recognize, multispace1))).take(),
     )
     .parse_next(input)
 }
@@ -35,6 +55,8 @@ pub fn multicomment1<'de>(input: &mut &'de BStr) -> PResult<&'de [u8]> {
 mod tests {
     use rstest::rstest;
 
+    use crate::extraction::extract;
+
     use super::*;
 
     #[rstest]
@@ -42,7 +64,7 @@ mod tests {
     #[case(b"% Another comment\r\n", b"Another comment")]
     fn extract_comment(#[case] input: &[u8], #[case] expected: &[u8]) {
         let mut input = input.as_ref();
-        let c = comment(&mut input).unwrap();
+        let Comment(c) = extract(&mut input).unwrap();
         assert!(input.is_empty(), "{:?}", input);
         assert_eq!(expected, c);
     }