+++ src/pdf/trailer_block/mod.rs
@@ -1,0 +1,11 @@
+mod block;
+mod plain;
+mod ref_location;
+mod startxref;
+mod stream;
+mod trailer;
+
+pub use block::XRefTrailerBlock;
+pub use ref_location::RefLocation;
+pub use startxref::StartXRef;
+pub use trailer::Trailer;