+++ src/pdf/trailer_block/plain.rs
@@ -1,3 +1,6 @@
+//! Describe how to extract "plain" references, i.e. indirect objects that are not compressed
+//! within an PDF [stream](crate::extraction::Stream).
+
 use std::{fmt::Debug, str::FromStr};
 
 use winnow::{
@@ -8,13 +11,11 @@ use winnow::{
     BStr, PResult, Parser,
 };
 
-use crate::{
-    extraction::{Extract, ReferenceId},
-    pdf::TrailerDict,
-};
+use crate::extraction::{extract, Extract, ReferenceId};
 
-use super::RefLocation;
+use super::{RefLocation, XRefTrailerBlock};
 
+/// Decimal number, with fixed number of characters.
 fn dec_num<'de, T, E>(count: usize) -> impl Parser<&'de BStr, T, ContextError>
 where
     T: FromStr<Err = E>,
@@ -29,6 +30,19 @@ where
     }
 }
 
+/// Extract a single cross-reference entry.
+///
+/// cross-reference entries are serialised using the following format:
+///
+/// ```raw
+/// nnnnnnnnnn ggggg n EOL
+/// ---------- ----- - ---
+///      |       |   |  |
+///      |       |   | 2-character end-of-line marker
+///      |       |  n/f: in use/free
+///      |     5-digit generation number
+///  10-digit byte offset of the indirect object
+/// ```
 fn xref_entry(input: &mut &BStr) -> PResult<(usize, u16, bool)> {
     trace("livre-ref-entry", move |i: &mut &BStr| {
         let (offset, gen) = separated_pair(dec_num(10), b' ', dec_num(5)).parse_next(i)?;
@@ -45,8 +59,16 @@ fn xref_entry(input: &mut &BStr) -> PResult<(usize, u16, bool)> {
     .parse_next(input)
 }
 
-fn xref_subsection<'de>(input: &mut &'de BStr) -> PResult<Vec<(ReferenceId, usize)>> {
+/// Extract a cross-reference subsection.
+///
+/// ```raw
+/// 0 6                  -> Id of the first object, and number of objects in the subsection
+/// 0000000003 65535 f   -> Cross-reference entry
+/// 0000000017 00000 n   -> Another Xref entry
+/// ```
+fn xref_subsection(input: &mut &BStr) -> PResult<Vec<(ReferenceId, usize)>> {
     let (initial, n) = separated_pair(usize::extract, b' ', usize::extract).parse_next(input)?;
+
     line_ending(input)?;
 
     let entries: Vec<(usize, u16, bool)> = repeat(n, xref_entry).parse_next(input)?;
@@ -65,27 +87,28 @@ fn xref_subsection<'de>(input: &mut &'de BStr) -> PResult<Vec<(ReferenceId, usiz
     Ok(res)
 }
 
-pub fn xref_sections(input: &mut &BStr) -> PResult<Vec<(ReferenceId, RefLocation)>> {
+/// Extract the full cross-reference table.
+pub fn xref(input: &mut &BStr) -> PResult<Vec<(ReferenceId, RefLocation)>> {
     (b"xref", multispace1).parse_next(input)?;
 
     let mut it = iterator(*input, terminated(xref_subsection, multispace0));
     let res = it
         .flatten()
-        .map(|(r, loc)| (r, RefLocation::Uncompressed(loc)))
+        .map(|(r, loc)| (r, RefLocation::Plain(loc)))
         .collect();
     *input = it.finish()?.0;
 
     Ok(res)
 }
 
-pub fn xref(input: &mut &BStr) -> PResult<(TrailerDict, Vec<(ReferenceId, RefLocation)>)> {
+pub fn block(input: &mut &BStr) -> PResult<XRefTrailerBlock> {
     trace("livre-xref-plain", move |i: &mut &BStr| {
-        let xrefs = xref_sections(i)?;
+        let xrefs = xref(i)?;
 
         (multispace0, b"trailer", multispace1).parse_next(i)?;
-        let dict = TrailerDict::extract(i)?;
+        let trailer = extract(i)?;
 
-        Ok((dict, xrefs))
+        Ok(XRefTrailerBlock { trailer, xrefs })
     })
     .parse_next(input)
 }
@@ -155,13 +178,13 @@ mod tests {
             0000000300 00000 n\r
         "},
         vec![
-            ((1, 0).into(), RefLocation::Uncompressed(200)),
-            ((2, 1).into(), RefLocation::Uncompressed(220)),
-            ((4, 0).into(), RefLocation::Uncompressed(300)),
+            ((1, 0).into(), RefLocation::Plain(200)),
+            ((2, 1).into(), RefLocation::Plain(220)),
+            ((4, 0).into(), RefLocation::Plain(300)),
         ]
     )]
     fn xref_extraction(#[case] input: &[u8], #[case] expected: Vec<(ReferenceId, RefLocation)>) {
-        let res = xref_sections(&mut input.as_ref()).unwrap();
+        let res = xref(&mut input.as_ref()).unwrap();
         assert_eq!(expected, res)
     }
 }